REF: split supervisor and worker (#279)
