+++ xinference/core/worker.py
@@ -14,195 +14,20 @@
 
 import asyncio
 import platform
-import time
-from dataclasses import dataclass
 from logging import getLogger
-from typing import Any, Callable, Dict, List, Optional, Set, Tuple
+from typing import Any, Dict, List, Optional, Set, Tuple
 
 import xoscar as xo
 
 from ..core import ModelActor
 from ..model.llm import LLMFamilyV1, LLMSpecV1
-from .resource import ResourceStatus, gather_node_info
+from .resource import gather_node_info
+from .utils import log_async, log_sync
 
 logger = getLogger(__name__)
 
 
-DEFAULT_NODE_DEAD_TIMEOUT = 30
-DEFAULT_NODE_CHECK_INTERVAL = 1
-
-
-def log(func: Callable):
-    # TODO: support non-async function
-    import time
-    from functools import wraps
-
-    @wraps(func)
-    async def wrapped(*args, **kwargs):
-        logger.debug(f"Enter {func.__name__}, args: {args}, kwargs: {kwargs}")
-        start = time.time()
-        ret = await func(*args, **kwargs)
-        logger.debug(
-            f"Leave {func.__name__}, elapsed time: {int(time.time() - start)} ms"
-        )
-        return ret
-
-    return wrapped
-
-
-@dataclass
-class WorkerStatus:
-    update_time: float
-    status: Dict[str, ResourceStatus]
-
-
-class SupervisorActor(xo.Actor):
-    def __init__(self):
-        super().__init__()
-        self._worker_address_to_worker: Dict[str, xo.ActorRefType[WorkerActor]] = {}
-        self._model_uid_to_worker: Dict[str, xo.ActorRefType[WorkerActor]] = {}
-        self._worker_status: Dict[str, WorkerStatus] = {}
-
-    @classmethod
-    def uid(cls) -> str:
-        return "supervisor"
-
-    async def __post_create__(self):
-        self._check_dead_nodes_task = asyncio.create_task(self._check_dead_nodes())
-
-    async def __pre_destroy__(self):
-        self._check_dead_nodes_task.cancel()
-
-    async def _choose_worker(self) -> xo.ActorRefType["WorkerActor"]:
-        # TODO: better allocation strategy.
-        min_running_model_count = None
-        target_worker = None
-        for worker in self._worker_address_to_worker.values():
-            running_model_count = await worker.get_model_count()
-            if (
-                min_running_model_count is None
-                or running_model_count < min_running_model_count
-            ):
-                min_running_model_count = running_model_count
-                target_worker = worker
-
-        if target_worker:
-            return target_worker
-
-        raise RuntimeError("No available worker found")
-
-    async def launch_builtin_model(
-        self,
-        model_uid: str,
-        model_name: str,
-        model_size_in_billions: Optional[int],
-        model_format: Optional[str],
-        quantization: Optional[str],
-        **kwargs,
-    ) -> xo.ActorRefType["ModelActor"]:
-        logger.debug(
-            (
-                f"Enter launch_builtin_model, model_uid: %s, model_name: %s, model_size: %s, "
-                f"model_format: %s, quantization: %s"
-            ),
-            model_uid,
-            model_name,
-            str(model_size_in_billions) if model_size_in_billions else "",
-            model_format,
-            quantization,
-        )
-
-        if model_uid in self._model_uid_to_worker:
-            raise ValueError(f"Model is already in the model list, uid: {model_uid}")
-
-        worker_ref = await self._choose_worker()
-        model_ref = yield worker_ref.launch_builtin_model(
-            model_uid=model_uid,
-            model_name=model_name,
-            model_size_in_billions=model_size_in_billions,
-            model_format=model_format,
-            quantization=quantization,
-            **kwargs,
-        )
-        # TODO: not protected.
-        self._model_uid_to_worker[model_uid] = worker_ref
-
-        raise xo.Return(model_ref)
-
-    async def _check_dead_nodes(self):
-        while True:
-            dead_nodes = []
-            for address, status in self._worker_status.items():
-                if time.time() - status.update_time > DEFAULT_NODE_DEAD_TIMEOUT:
-                    dead_models = []
-                    for model_uid in self._model_uid_to_worker:
-                        if self._model_uid_to_worker[model_uid].address == address:
-                            dead_models.append(model_uid)
-                    logger.error(
-                        "Worker timeout. address: %s, influenced models: %s",
-                        address,
-                        dead_models,
-                    )
-                    dead_nodes.append(address)
-
-            for address in dead_nodes:
-                self._worker_status.pop(address)
-                self._worker_address_to_worker.pop(address)
-            await asyncio.sleep(5)
-
-    @log
-    async def terminate_model(self, model_uid: str):
-        if model_uid not in self._model_uid_to_worker:
-            raise ValueError(f"Model not found in the model list, uid: {model_uid}")
-
-        worker_ref = self._model_uid_to_worker[model_uid]
-        await worker_ref.terminate_model(model_uid=model_uid)
-        del self._model_uid_to_worker[model_uid]
-
-    @log
-    async def get_model(self, model_uid: str) -> xo.ActorRefType["ModelActor"]:
-        if model_uid not in self._model_uid_to_worker:
-            raise ValueError(f"Model not found in the model list, uid: {model_uid}")
-
-        worker_ref = self._model_uid_to_worker[model_uid]
-        return await worker_ref.get_model(model_uid=model_uid)
-
-    @log
-    async def describe_model(self, model_uid: str) -> Dict[str, Any]:
-        if model_uid not in self._model_uid_to_worker:
-            raise ValueError(f"Model not found in the model list, uid: {model_uid}")
-
-        worker_ref = self._model_uid_to_worker[model_uid]
-        return await worker_ref.describe_model(model_uid=model_uid)
-
-    @log
-    async def list_models(self) -> Dict[str, Dict[str, Any]]:
-        ret = {}
-        for worker in self._worker_address_to_worker.values():
-            ret.update(await worker.list_models())
-        return ret
-
-    def is_local_deployment(self) -> bool:
-        # TODO: temporary.
-        return (
-            len(self._worker_address_to_worker) == 1
-            and list(self._worker_address_to_worker)[0] == self.address
-        )
-
-    @log
-    async def add_worker(self, worker_address: str):
-        assert worker_address not in self._worker_address_to_worker
-
-        worker_ref = await xo.actor_ref(address=worker_address, uid=WorkerActor.uid())
-        self._worker_address_to_worker[worker_address] = worker_ref
-        logger.info("Worker %s has been added successfully", worker_address)
-
-    async def report_worker_status(
-        self, worker_address: str, status: Dict[str, ResourceStatus]
-    ):
-        self._worker_status[worker_address] = WorkerStatus(
-            update_time=time.time(), status=status
-        )
+DEFAULT_NODE_HEARTBEAT_INTERVAL = 1
 
 
 class WorkerActor(xo.Actor):
@@ -224,6 +49,8 @@ class WorkerActor(xo.Actor):
         return "worker"
 
     async def __post_create__(self):
+        from .supervisor import SupervisorActor
+
         self._supervisor_ref: xo.ActorRefType["SupervisorActor"] = await xo.actor_ref(
             address=self._supervisor_address, uid=SupervisorActor.uid()
         )
@@ -233,7 +60,8 @@ class WorkerActor(xo.Actor):
     async def __pre_destroy__(self):
         self._upload_task.cancel()
 
-    async def get_model_count(self) -> int:
+    @log_sync
+    def get_model_count(self) -> int:
         return len(self._model_uid_to_model)
 
     def _choose_subpool(self) -> str:
@@ -282,7 +110,7 @@ class WorkerActor(xo.Actor):
             "quantization": quantization,
         }
 
-    @log
+    @log_async(logger=logger)
     async def launch_builtin_model(
         self,
         model_uid: str,
@@ -337,7 +165,7 @@ class WorkerActor(xo.Actor):
         self._subpool_address_to_model_uids[subpool_address].add(model_uid)
         return model_ref
 
-    @log
+    @log_async(logger=logger)
     async def terminate_model(self, model_uid: str):
         if model_uid not in self._model_uid_to_model:
             raise ValueError(f"Model not found in the model list, uid: {model_uid}")
@@ -351,22 +179,22 @@ class WorkerActor(xo.Actor):
             if model_uid in self._subpool_address_to_model_uids[subpool_address]:
                 self._subpool_address_to_model_uids[subpool_address].remove(model_uid)
 
-    @log
-    async def list_models(self) -> Dict[str, Dict[str, Any]]:
+    @log_sync(logger=logger)
+    def list_models(self) -> Dict[str, Dict[str, Any]]:
         ret = {}
         for k, v in self._model_uid_to_model_spec.items():
             ret[k] = self._to_llm_description(v[0], v[1], v[2])
         return ret
 
-    @log
-    async def get_model(self, model_uid: str) -> xo.ActorRefType["ModelActor"]:
+    @log_sync(logger=logger)
+    def get_model(self, model_uid: str) -> xo.ActorRefType["ModelActor"]:
         if model_uid not in self._model_uid_to_model:
             raise ValueError(f"Model not found in the model list, uid: {model_uid}")
 
         return self._model_uid_to_model[model_uid]
 
-    @log
-    async def describe_model(self, model_uid: str) -> Dict[str, Any]:
+    @log_sync(logger=logger)
+    def describe_model(self, model_uid: str) -> Dict[str, Any]:
         if model_uid not in self._model_uid_to_model:
             raise ValueError(f"Model not found in the model list, uid: {model_uid}")
 
@@ -393,6 +221,6 @@ class WorkerActor(xo.Actor):
             ) as ex:  # pragma: no cover  # noqa: E722  # nosec  # pylint: disable=bare-except
                 logger.error(f"Failed to upload node info: {ex}")
             try:
-                await asyncio.sleep(DEFAULT_NODE_CHECK_INTERVAL)
+                await asyncio.sleep(DEFAULT_NODE_HEARTBEAT_INTERVAL)
             except asyncio.CancelledError:  # pragma: no cover
                 break
